backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.4.1
httpx>=0.24.0
python-dotenv==1.0.0
reportlab==4.0.7

